refactor: move platform-specific code into lib/macos/ and lib/linux/ folders
- lib/macos/launchd-paths.ts — moved from lib/launchd-paths.ts
- lib/macos/plist-generate.ts — moved from lib/plist-generate.ts
- lib/macos/installer.ts — macOS-only installer functions (launchctl, plist, LAUNCH_AGENTS_DIR)
- lib/linux/in-process-scheduler.ts — moved from chatbot/a2a/lib/
- lib/linux/__tests__/in-process-scheduler.test.ts — moved alongside implementation

Old paths kept as barrel re-exports for backward compat.
vitest.node.config.ts gains resolve aliases so @ and @@ work in lib/linux/ tests.
